refactor: #64 - 전체적인 리팩토링
count query 반환 타입 int -> Long
Entity 식별자 변경 LikeInform -> Like
메소드 식별자 수정

+++ src/main/java/study/board/repository/BaseCommentRepository.java
@@ -8,6 +8,6 @@ import study.board.entity.comment.BaseComment;
 public interface BaseCommentRepository extends JpaRepository<BaseComment, Long> {
 
     @Query("select count(*) from BaseComment bc where bc.post.id = :postId and bc.isRemoved = false")
-    int countByPost_Id(@Param("postId") Long postId);
+    Long countByPost_Id(@Param("postId") Long postId);
 
 }
